test(core,react): split large test files into focused modules
Split 5 oversized test files into 19 smaller, focused files:

- normal-mode.test.ts (1,969L) → 6 files by functional area
- motions.test.ts (1,069L) → 4 files by motion type
- motion-resolver.test.ts (833L) → 3 files by resolution category
- visual-mode.test.ts (733L) → 3 files by mode variant
- useVim.test.tsx (1,100L) → 3 files by feature group

All 1,067 tests pass with unchanged coverage.
