+++ packages/core/src/__tests__/motion-resolver-file-screen.test.ts
@@ -3,17 +3,14 @@ import { TextBuffer } from "../buffer";
 import { resolveMotion } from "../motion-resolver";
 import type { VimContext } from "../types";
 
-// Helper: create a TextBuffer from an array of lines
 function buf(lines: string[]): TextBuffer {
   return new TextBuffer(lines.join("\n"));
 }
 
-// Helper: concisely create a cursor position
 function cur(line: number, col: number) {
   return { line, col };
 }
 
-// Helper: create a minimal VimContext for viewport-dependent motions
 function createCtx(overrides: Partial<VimContext> = {}): VimContext {
   return {
     mode: "normal",
@@ -47,366 +44,6 @@ function createCtx(overrides: Partial<VimContext> = {}): VimContext {
     ...overrides,
   };
 }
-
-// ---------- Basic movement: h / ArrowLeft ----------
-
-describe("resolveMotion: h and ArrowLeft", () => {
-  it("h moves cursor left by count", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("h", cur(0, 3), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 2));
-  });
-
-  it("h moves left by count=2", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("h", cur(0, 4), b, 2, false);
-    expect(result!.cursor).toEqual(cur(0, 2));
-  });
-
-  it("h clamps at column 0", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("h", cur(0, 1), b, 10, false);
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-
-  it("ArrowLeft behaves the same as h", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("ArrowLeft", cur(0, 3), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 2));
-  });
-
-  it("ArrowLeft clamps at column 0", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("ArrowLeft", cur(0, 0), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-});
-
-// ---------- Basic movement: j / ArrowDown ----------
-
-describe("resolveMotion: j and ArrowDown", () => {
-  it("j moves cursor down by count", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("j", cur(0, 2), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(1, 2));
-  });
-
-  it("j clamps at last line", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("j", cur(0, 0), b, 10, false);
-    expect(result!.cursor.line).toBe(1);
-  });
-
-  it("j clamps col when target line is shorter", () => {
-    const b = buf(["hello world", "hi"]);
-    const result = resolveMotion("j", cur(0, 8), b, 1, false);
-    expect(result!.cursor).toEqual(cur(1, 1));
-  });
-
-  it("j returns linewise range", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("j", cur(0, 0), b, 1, false);
-    expect(result!.range.linewise).toBe(true);
-  });
-
-  it("ArrowDown behaves the same as j", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("ArrowDown", cur(0, 2), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(1, 2));
-  });
-});
-
-// ---------- Basic movement: k / ArrowUp ----------
-
-describe("resolveMotion: k and ArrowUp", () => {
-  it("k moves cursor up by count", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("k", cur(1, 2), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 2));
-  });
-
-  it("k clamps at first line", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("k", cur(1, 0), b, 10, false);
-    expect(result!.cursor.line).toBe(0);
-  });
-
-  it("k clamps col when target line is shorter", () => {
-    const b = buf(["hi", "hello world"]);
-    const result = resolveMotion("k", cur(1, 8), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 1));
-  });
-
-  it("k returns linewise range", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("k", cur(1, 0), b, 1, false);
-    expect(result!.range.linewise).toBe(true);
-  });
-
-  it("ArrowUp behaves the same as k", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("ArrowUp", cur(1, 2), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 2));
-  });
-});
-
-// ---------- Basic movement: l / ArrowRight ----------
-
-describe("resolveMotion: l and ArrowRight", () => {
-  it("l moves cursor right by count", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("l", cur(0, 0), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 1));
-  });
-
-  it("l clamps at end of line", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("l", cur(0, 3), b, 10, false);
-    expect(result!.cursor).toEqual(cur(0, 4));
-  });
-
-  it("l does not move on empty line", () => {
-    const b = buf([""]);
-    const result = resolveMotion("l", cur(0, 0), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-
-  it("ArrowRight behaves the same as l", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("ArrowRight", cur(0, 0), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 1));
-  });
-
-  it("ArrowRight clamps at end of line", () => {
-    const b = buf(["hi"]);
-    const result = resolveMotion("ArrowRight", cur(0, 1), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 1));
-  });
-});
-
-// ---------- Word motions: w ----------
-
-describe("resolveMotion: w (word forward)", () => {
-  it("moves to the beginning of the next word", () => {
-    const b = buf(["hello world"]);
-    const result = resolveMotion("w", cur(0, 0), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 6));
-  });
-
-  it("treats punctuation as a separate word", () => {
-    const b = buf(["foo.bar"]);
-    const result = resolveMotion("w", cur(0, 0), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 3));
-  });
-
-  it("moves across lines", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("w", cur(0, 0), b, 2, false);
-    expect(result!.cursor).toEqual(cur(1, 0));
-  });
-
-  it("returns non-linewise, non-inclusive range", () => {
-    const b = buf(["hello world"]);
-    const result = resolveMotion("w", cur(0, 0), b, 1, false);
-    expect(result!.range.linewise).toBe(false);
-    expect(result!.range.inclusive).toBe(false);
-  });
-});
-
-// ---------- Word motions: e ----------
-
-describe("resolveMotion: e (word end)", () => {
-  it("moves to the end of the current word", () => {
-    const b = buf(["hello world"]);
-    const result = resolveMotion("e", cur(0, 0), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 4));
-  });
-
-  it("moves to the end of the next word when at word end", () => {
-    const b = buf(["hello world"]);
-    const result = resolveMotion("e", cur(0, 4), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 10));
-  });
-
-  it("moves across lines", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("e", cur(0, 4), b, 1, false);
-    expect(result!.cursor).toEqual(cur(1, 4));
-  });
-
-  it("returns inclusive range", () => {
-    const b = buf(["hello world"]);
-    const result = resolveMotion("e", cur(0, 0), b, 1, false);
-    expect(result!.range.inclusive).toBe(true);
-  });
-});
-
-// ---------- Word motions: b ----------
-
-describe("resolveMotion: b (word backward)", () => {
-  it("moves to the beginning of the previous word", () => {
-    const b = buf(["hello world"]);
-    const result = resolveMotion("b", cur(0, 8), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 6));
-  });
-
-  it("moves backward across lines", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("b", cur(1, 0), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-
-  it("does not go past the beginning", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("b", cur(0, 0), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-});
-
-// ---------- Word motions: W (WORD forward) ----------
-
-describe("resolveMotion: W (WORD forward)", () => {
-  it("treats punctuation as part of the WORD", () => {
-    const b = buf(["foo.bar baz"]);
-    const result = resolveMotion("W", cur(0, 0), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 8));
-  });
-
-  it("moves across lines", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("W", cur(0, 0), b, 1, false);
-    expect(result!.cursor).toEqual(cur(1, 0));
-  });
-
-  it("stays at end of file when no more WORDs", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("W", cur(0, 0), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-});
-
-// ---------- Word motions: B (WORD backward) ----------
-
-describe("resolveMotion: B (WORD backward)", () => {
-  it("treats punctuation as part of the WORD backwards", () => {
-    const b = buf(["foo.bar baz"]);
-    const result = resolveMotion("B", cur(0, 8), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-
-  it("moves backward across lines", () => {
-    const b = buf(["hello", "world"]);
-    const result = resolveMotion("B", cur(1, 0), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-
-  it("stays at beginning of file", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("B", cur(0, 0), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-});
-
-// ---------- Line motions: 0 ----------
-
-describe("resolveMotion: 0 (beginning of line)", () => {
-  it("moves to column 0", () => {
-    const b = buf(["  hello"]);
-    const result = resolveMotion("0", cur(0, 5), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-
-  it("does nothing when already at column 0", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("0", cur(0, 0), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-
-  it("range is non-linewise, non-inclusive", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("0", cur(0, 3), b, 1, false);
-    expect(result!.range.linewise).toBe(false);
-    expect(result!.range.inclusive).toBe(false);
-  });
-});
-
-// ---------- Line motions: ^ ----------
-
-describe("resolveMotion: ^ (first non-blank)", () => {
-  it("moves to the first non-whitespace character", () => {
-    const b = buf(["  hello"]);
-    const result = resolveMotion("^", cur(0, 5), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 2));
-  });
-
-  it("moves to column 0 when line has no leading whitespace", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("^", cur(0, 3), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-
-  it("moves to column 0 on an empty line", () => {
-    const b = buf([""]);
-    const result = resolveMotion("^", cur(0, 0), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-
-  it("range is inclusive", () => {
-    const b = buf(["  hello"]);
-    const result = resolveMotion("^", cur(0, 5), b, 1, false);
-    expect(result!.range.inclusive).toBe(true);
-  });
-});
-
-// ---------- Line motions: $ ----------
-
-describe("resolveMotion: $ (end of line)", () => {
-  it("moves to the last character of the line", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("$", cur(0, 0), b, 1, false);
-    expect(result).not.toBeNull();
-    expect(result!.cursor).toEqual(cur(0, 4));
-  });
-
-  it("stays at col=0 on an empty line", () => {
-    const b = buf([""]);
-    const result = resolveMotion("$", cur(0, 0), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 0));
-  });
-
-  it("does nothing when already at end of line", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("$", cur(0, 4), b, 1, false);
-    expect(result!.cursor).toEqual(cur(0, 4));
-  });
-
-  it("range is inclusive", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("$", cur(0, 0), b, 1, false);
-    expect(result!.range.inclusive).toBe(true);
-  });
-
-  it("range end is at lineLength (exclusive end for operators)", () => {
-    const b = buf(["hello"]);
-    const result = resolveMotion("$", cur(0, 1), b, 1, false);
-    expect(result!.range.end).toEqual(cur(0, 5));
-  });
-});
 
 // ---------- File motions: G ----------
